flask_restplus/requirements_back.txt
absl-py==0.1.12
aniso8601==3.0.0
astor==0.6.2
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
Django==2.0.3
Flask==0.12.2
flask-restplus==0.10.1
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
image==1.5.20
imutils==0.4.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Keras==2.1.5
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mysqlclient==1.3.12
nltk==3.2.5
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
pandas-profiling==1.4.0
Pillow==5.0.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.5
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
textblob==0.15.1
urllib3==1.22
Werkzeug==0.14.1
